Changed: internal scrollbar precision in IupCanvas and IupMatrix from float to double.
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@4000 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/iup_drvinfo.h
@@ -50,7 +50,7 @@ int iupdrvGetScreenDepth(void);
 
 /** Retrieve the default desktop resolution in dpi (dots or pixels per inch).
  * \ingroup drvinfo */
-float iupdrvGetScreenDpi(void);
+double iupdrvGetScreenDpi(void);
 
 /** Returns a string with the system version number.
  * \ingroup drvinfo */
